#953: Move command property value providers to core API
